requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.4
django-bootstrap-datepicker-plus==3.0.5
django-cors-headers==3.7.0
django-filter==2.4.0
django-import-export==2.5.0
django-oauth-toolkit==1.4.0
djangorestframework==3.12.2
djangorestframework-oauth==1.1.0
djangorestframework-simplejwt==4.6.0
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
MarkupPy==1.14
mysqlclient==2.0.2
numpy==1.19.5
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.2.1
provider==0.1.0
Pygments==2.8.0
PyJWT==2.0.1
python-dateutil==2.8.1
pytz==2020.4
PyYAML==5.3.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.3
xlrd==2.0.1
xlwt==1.3.0
